Move class to only package where it is used
