+++ subprojects/core/src/main/java/org/gradle/api/internal/file/archive/compression/URIBuilder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.resources;
+package org.gradle.api.internal.file.archive.compression;
 
 import org.gradle.util.GUtil;
